feat(stage-layout): update stage
