--- src/client/layouts/on-stage.html
@@ -1,1 +1,0 @@
-TODO ...
