+++ src/client/layouts/on-stage/index.css
@@ -1,0 +1,6 @@
+#stageFrame {
+    width: 100%;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
