+++ src/client/layouts/on-stage/on-stage.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+        <link rel="stylesheet" href="../../../../node_modules/bulma/css/bulma.min.css" />
+        <link rel="stylesheet" href="index.css" />
+        <script defer src="../../../../node_modules/@fortawesome/fontawesome-free/js/all.min.js"></script>
+        <title>On stage</title>
+    </head>
+    <body>
+        <section class="section" id="form">
+            <div class="container">
+                <h1 class="title">
+                    Talk control
+                </h1>
+                <p class="subtitle">
+                    Take control of your presentation by entering its url below
+                </p>
+                <!-- Enter presentation url form area -->
+                <div class="columns">
+                    <div class="column is-half">
+                        <div class="columns">
+                            <div class="column is-four-fifths">
+                                <input
+                                    class="input"
+                                    type="url"
+                                    placeholder="Enter the url of your presentation : http://my.presentation.url/index.html"
+                                    id="inputPresentation"
+                                    value="http://127.0.0.1:5000"
+                                />
+                            </div>
+                            <div class="column">
+                                <button id="btnValidate" class="button">Validate Url</button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+
+                <!-- Error message area -->
+                <div class="columns">
+                    <div class="column is-half">
+                        <div id="urlError" class="notification is-danger is-hidden">
+                            <button class="delete"></button>
+                            URL is not valid
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </section>
+        <section>
+            <iframe id="stageFrame" class="is-hidden"></iframe>
+        </section>
+        <script src="./index.js"></script>
+    </body>
+</html>
