--- src/client/talk-control-master/index.css
@@ -1,6 +1,0 @@
-#stageFrame {
-    width: 100%;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
